Create BIMOB_Login_Page
This is the updated code for the login page of BIMOB. First, the responsible person for creating the BIMOB account will create and provide the username & password. Then the user has to log in to the BIMOB login portal. This is why, create account section is missing here. 
First, the user will see the login button. After clicking the login button, the user will see the username and password textboxes appear on the screen. After 10 seconds of inactivity, the username and password textboxes will disappear again, and the login button will be visible only.